Factor out prompt creation from actor document.
