+++ module/applications/global/_module.mjs
@@ -1,2 +1,3 @@
 export { default as RollPrompt } from "./roll-prompt.mjs";
 export { default as LegendPointHistoryEarnedPrompt } from "./lp-history.mjs";
+export { default as PromptFactory } from "./prompt-factory.mjs";
