+++ module/documents/actor.mjs
@@ -11,6 +11,7 @@ import LpTrackingData from "../data/advancement/lp-tracking.mjs";
 import { sum } from "../utils.mjs";
 import KnockDownItemsPrompt from "../applications/global/knock-down-prompt.mjs";
 import JumpUpItemsPrompt from "../applications/global/jump-up-prompt.mjs";
+import PromptFactory from "../applications/global/prompt-factory.mjs";
 
 /**
  * Extend the base Actor class to implement additional system-specific logic.
@@ -18,10 +19,7 @@ import JumpUpItemsPrompt from "../applications/global/jump-up-prompt.mjs";
  */
 export default class ActorEd extends Actor {
 
-  /** @inheritDoc */
-  static async createDialog( data = {}, { parent = null, pack = null, ...options } = {} ) {
-    return DocumentCreateDialog.waitPrompt( data, { documentCls: Actor, parent, pack, options } );
-  }
+  _promptFactory = new PromptFactory( this );
 
   /**
    * Returns the namegiver item if this actor has one (has to be of type "character" or "npc" for this).
@@ -29,6 +27,11 @@ export default class ActorEd extends Actor {
    */
   get namegiver() {
     return this.items.filter( item => item.type === 'namegiver' )[0];
+  }
+
+  /** @inheritDoc */
+  static async createDialog( data = {}, { parent = null, pack = null, ...options } = {} ) {
+    return DocumentCreateDialog.waitPrompt( data, { documentCls: Actor, parent, pack, options } );
   }
 
   /**
@@ -93,7 +96,10 @@ export default class ActorEd extends Actor {
    */
   async legendPointHistoryEarned() {
     // let history = await getLegendPointHistoryData( actor );
-    const lpUpdateData = await LegendPointHistoryEarnedPrompt.waitPrompt( new LpTrackingData( this.system.lp.toObject() ), { actor: this } );
+    const lpUpdateData = await LegendPointHistoryEarnedPrompt.waitPrompt(
+      new LpTrackingData( this.system.lp.toObject() ),
+      { actor: this }
+    );
     return this.update( { system: { lp: lpUpdateData } } );
   }
 
@@ -279,7 +285,7 @@ export default class ActorEd extends Actor {
       sourceActor: this.name,
       step: recoveryStep
     } );
-    const recoveryFinalStep = { base: recoveryStep }
+    const recoveryFinalStep = { base: recoveryStep };
     const edRollOptions = EdRollOptions.fromActor(
       {
         testType: "effect",
@@ -612,7 +618,6 @@ export default class ActorEd extends Actor {
       }
 
       return changes.concat(
-
         e.changes.map( ( c ) => {
           // eslint-disable-next-line no-param-reassign
           c = foundry.utils.duplicate( c );
@@ -763,141 +768,17 @@ export default class ActorEd extends Actor {
     return this.updateEmbeddedDocuments( "Item", updates );
   }
 
-  async _getPrompt( promptType ) {
-    const DialogClass = foundry.applications.api.DialogV2;
-    switch ( promptType ) {
-      case "recovery":
-        const buttons = [];
-        if ( this.system.characteristics.recoveryTestsRecource.value > 0 ) buttons.push( {
-          action: "recovery",
-          label: game.i18n.localize( "ED.Dialogs.Buttons.recovery"),
-          icon: "fa-light fa-heart-circle-plus",
-          class: "recovery default button-recovery",
-          default: false,
-          callback: _ => { return "recovery" },
-        } );
-        if ( this.system.characteristics.recoveryTestsRecource.stunRecoveryAvailabilty ) buttons.push( {
-          action: "recoverStun",
-          label: game.i18n.localize( "ED.Dialogs.Buttons.recoverStun"),
-          icon: "fa-light fa-head-side-medical",
-          class: "recoverStun default button-recoverStun",
-          default: false,
-          callback: _ => { return "recoverStun" },
-        } );
-        buttons.push( {
-          action: "nightRest",
-          label: game.i18n.localize( "ED.Dialogs.Buttons.nightRest"),
-          icon: "fa-duotone fa-campfire",
-          class: "nightRest default button-nightRest",
-          default: true,
-          callback: _ => { return "nightRest" },
-        } );
-        buttons.push( {
-          action: "close",
-          label: game.i18n.localize( "ED.Dialogs.Buttons.cancel"),
-          icon: "fa-light fa-times",
-          class: "cancel default button-cancel",
-          default: false,
-        } );
-
-        return DialogClass.wait( {
-          id: "recovery-mode-prompt",
-          uniqueId: String( ++globalThis._appId ),
-          classes: ["earthdawn4e", "recovery-prompt"],
-          window: {
-            title: game.i18n.localize( "ED.Dialogs.Title.recovery" ),
-            minimizable: false,
-          },
-          modal: false,
-          buttons: buttons,
-        } );
-      case "takeDamage":
-        const fields = foundry.data.fields;
-        const formFields = {
-          damage: new fields.NumberField( {
-            required: true,
-            name: "damage",
-            initial: 1,
-            integer: true,
-            positive: true,
-            label: "ED.Dialogs.damage",
-            hint: "localize: The amount of damage to take",
-          } ),
-          damageType: new fields.StringField( {
-            required: true,
-            nullable: false,
-            name: "damageType",
-            initial: "standard",
-            blank: false,
-            label: "ED.Dialogs.damageType",
-            hint: "localize: The type of damage to take",
-            choices: {
-              standard: "ED.Dialogs.damageStandard",
-              stun: "ED.Dialogs.damageStun",
-            },
-          } ),
-          armorType: new fields.StringField( {
-            required: true,
-            nullable: false,
-            name: "armorType",
-            initial: "physical",
-            blank: false,
-            label: "ED.Dialogs.armorType",
-            hint: "localize: The type of armor to use",
-            choices: {
-              physical: "ED.Dialogs.physical",
-              mystical: "ED.Dialogs.mystical",
-            },
-          } ),
-          ignoreArmor: new fields.BooleanField( {
-            required: true,
-            name: "ignoreArmor",
-            initial: false,
-            label: "ED.Dialogs.ignoreArmor",
-            hint: "localize: Whether to ignore armor when taking damage",
-          } ),
-        };
-        return DialogClass.wait( {
-          id: "take-damage-prompt",
-          uniqueId: String( ++globalThis._appId ),
-          classes: ["earthdawn4e", "take-damage-prompt", "take-damage__dialog"],
-          //tag: "form",
-          window: {
-            title: game.i18n.localize( "ED.Dialogs.Title.takeDamage" ),
-            minimizable: false,
-          },
-          form: {
-            submitOnChange: false,
-            closeOnSubmit: true,
-          },
-          modal: false,
-          buttons: [
-            {
-              action: "takeDamage",
-              label: game.i18n.localize( "ED.Dialogs.Buttons.takeDamage"),
-              icon: "fa-solid fa-heart-crack",
-              class: "takeDamage default button__take-damage",
-              default: false,
-              callback: ( event, button, dialog ) => {
-                const formData = new FormDataExtended( button.form );
-                return formData.object;
-              }
-            },
-            {
-              action: "close",
-              label: game.i18n.localize( "ED.Dialogs.Buttons.cancel"),
-              icon: "fa-light fa-times",
-              class: "cancel default button-cancel",
-              default: true,
-            },
-          ],
-          content: await renderTemplate(
-            "systems/ed4e/templates/actor/prompts/take-damage-prompt.hbs",
-            formFields,
-          ),
-          rejectClose: false,
-        } );
-    }
+  /**
+   * Retrieves a specific prompt based on the provided prompt type.
+   * This method delegates the call to the `_promptFactory` instance's `getPrompt` method,
+   * effectively acting as a proxy to access various prompts defined within the factory.
+   *
+   * @param {( 'recovery' | 'takeDamage' )} promptType - The type of prompt to retrieve.
+   * @returns {Promise<any>} - A promise that resolves to the specific prompt instance or logic
+   * associated with the given `promptType`. The exact return type depends on promptType.
+   */
+  async getPrompt( promptType ) {
+    return this._promptFactory.getPrompt( promptType );
   }
 
 }
